Move content to new publishedproject.py module
